Moved all java files outside of helloWorld, as it is no longer necessary
