--- helloWorld/src/helloworld/HelloWorld.java
@@ -1,28 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package helloworld;
-
-/**
- *
- * @author faithima
- */
-public class HelloWorld {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        // TODO code application logic here
-        System.out.println("Hello world");
-        System.out.println("My name is Chase Glasper");
-        System.out.println("This is Robert");
-        System.out.println("This is Faith");
-        System.out.println("This is Owen King typing into the hello world program \n" +
-                "I plan on creating the executable for this program, and explaining in the readMe \n" +
-                "How it is created and how to build it yourself");
-    }
-    
-}
