+++ src/main/java/main.java
@@ -1,0 +1,10 @@
+package main.java;
+
+public class main {
+
+    public static void main(String [] args){
+        System.out.println("This is to test if I can make this executable");
+        Battler person1 = new Character();
+        Battler person2 = new Enemy();
+    }
+}
